Refactored some of the code to allow animation to be its own module
