+++ src/Camera.java
@@ -1,15 +1,13 @@
 import javafx.scene.image.Image;
 import org.opencv.core.CvType;
 import org.opencv.core.Mat;
+import org.opencv.core.Rect;
 import org.opencv.core.Scalar;
 import org.opencv.imgproc.Imgproc;
 import org.opencv.videoio.VideoCapture;
 
-/**
- * Created by Andrew Haworth.
- */
 public class Camera extends VideoOutput {
-    private Mat gray = new Mat(), thresholdImg = new Mat(), hsv = new Mat();
+    private Mat gray = new Mat(), thresholdImg = new Mat(), hsv = new Mat(), frame = new Mat();
 
     public Image standard(VideoCapture capture) {
         return createImage(duplicate(removeBackground(grabFrame(capture))));
@@ -23,5 +21,24 @@ public class Camera extends VideoOutput {
         Mat foreground = new Mat(frame.size(), CvType.CV_8UC3, new Scalar(0, 0, 0));
         frame.copyTo(foreground, thresholdImg);
         return foreground;
+    }
+
+    public Mat grabFrame(VideoCapture capture) {
+        if (capture.isOpened())
+            capture.read(frame);
+        else
+            System.out.println("No camera!");
+        if (!frame.empty()) {
+            if (frame.rows() <= frame.cols())
+                frame = new Mat(frame, new Rect((frame.cols() - frame.rows()) / 2, 0,
+                        frame.rows(), frame.rows()));
+            else
+                frame = new Mat(frame, new Rect(0, (frame.rows() - frame.cols()) / 2,
+                        frame.cols(), frame.cols()));
+            return frame;
+        } else {
+            System.out.println("Oh no! The frame is empty. Sending blank.");
+            return new Mat(1920, 1080, CvType.CV_8U, new Scalar(128, 128, 128));
+        }
     }
 }
